feat: move shim server side into an own module
